+++ demos/demo/src/androidTest/java/com/growingio/autotest/autotracker/webservices/CircleServiceTest.java
@@ -66,12 +66,12 @@ public class CircleServiceTest extends WebServicesTest {
     @BeforeAppOnCreate
     public static void beforeAppOnCreate() {
         DataHelper.deleteEventsDatabase();
-        DemoApplication.setConfiguration(TestTrackConfiguration.getTestConfig("growing.ff8a70351000af43"));
+        DemoApplication.setConfiguration(TestTrackConfiguration.getTestConfig("growing.bd71d91eb56f5f53"));
     }
 
     @Test
     public void circleServiceTest() {
-        String uri = "growing.ff8a70351000af43://growingio/webservice?serviceType=circle&wsUrl=" + Uri.encode(getWsUrl());
+        String uri = "growing.bd71d91eb56f5f53://growingio/webservice?serviceType=circle&wsUrl=" + Uri.encode(getWsUrl());
         Intent intent = new Intent();
         intent.setData(Uri.parse(uri));
         ActivityScenario.launch(intent);
