+++ demos/demo/src/androidTest/java/com/growingio/autotest/autotracker/webservices/DebuggerServiceTest.java
@@ -68,7 +68,7 @@ public class DebuggerServiceTest {
     @BeforeAppOnCreate
     public static void beforeAppOnCreate() {
         DataHelper.deleteEventsDatabase();
-        DemoApplication.setConfiguration(TestTrackConfiguration.getTestConfig("growing.ff8a70351000af43"));
+        DemoApplication.setConfiguration(TestTrackConfiguration.getTestConfig("growing.bd71d91eb56f5f53"));
     }
 
     protected String getWsUrl() {
@@ -126,7 +126,7 @@ public class DebuggerServiceTest {
         });
         mMockWebServer.enqueue(ready);
 
-        String uri = "growing.ff8a70351000af43://growingio/webservice?serviceType=debugger&wsUrl=" + Uri.encode(getWsUrl());
+        String uri = "growing.bd71d91eb56f5f53://growingio/webservice?serviceType=debugger&wsUrl=" + Uri.encode(getWsUrl());
         Intent intent = new Intent();
         intent.setData(Uri.parse(uri));
         ActivityScenario.launch(intent);
