+++ growingio-webservice/debugger/src/test/java/com/growingio/android/debugger/ScreenShotTest.java
@@ -58,7 +58,6 @@ public class ScreenShotTest {
             Truth.assertThat(screenshot).isNotNull();
         };
         ScreenshotProvider.get().registerScreenshotRefreshedListener(dispatchListener);
-        ScreenshotProvider.get().refreshScreenshot();
 
         ScreenshotProvider.get().unregisterScreenshotRefreshedListener();
     }
